+++ test/httpoison_test.exs
@@ -144,9 +144,13 @@ defmodule HTTPoisonTest do
   end
 
   test "put without body" do
-    assert_response(HTTPoison.put("localhost:4002/put"), fn response ->
+    response =
+      HTTPoison.put("localhost:4002/put", "", [{"Content-Type", "application/octet-stream"}])
+
+    assert_response(response, fn response ->
       assert Request.to_curl(response.request) ==
-               {:ok, "curl -X PUT http://localhost:4002/put"}
+               {:ok,
+                "curl -X PUT -H 'Content-Type: application/octet-stream' http://localhost:4002/put"}
     end)
   end
 
@@ -299,9 +303,6 @@ defmodule HTTPoisonTest do
       HTTPoison.get("localhost:4002/get", [], stream_to: self(), async: :once)
 
     assert_receive %HTTPoison.AsyncStatus{id: ^id, code: 200}, 100
-
-    refute_receive %HTTPoison.AsyncHeaders{id: ^id, headers: _headers}, 100
-    {:ok, ^resp} = HTTPoison.stream_next(resp)
     assert_receive %HTTPoison.AsyncHeaders{id: ^id, headers: headers}, 100
 
     refute_receive %HTTPoison.AsyncChunk{id: ^id, chunk: _chunk}, 100
@@ -355,22 +356,12 @@ defmodule HTTPoisonTest do
   end
 
   test "max_body_length limits body size" do
-    {:ok, socket} = :gen_tcp.listen(0, [:binary, {:active, false}, {:packet, :raw}])
-    {:ok, [buffer: buffer_size]} = :inet.getopts(socket, [:buffer])
-    :ok = :gen_tcp.close(socket)
-
-    max_length = Kernel.trunc(buffer_size * 1.5)
-
-    expected_length =
-      Float.ceil(max_length / buffer_size)
-      |> Kernel.*(buffer_size)
-      |> Kernel.trunc()
+    max_length = 64
 
     resp = HTTPoison.get("localhost:4002/stream/20", [], max_body_length: max_length)
 
     assert_response(resp, fn response ->
-      assert byte_size(response.body) <= expected_length
-      assert byte_size(response.body) >= max_length
+      assert byte_size(response.body) == max_length
     end)
   end
 
@@ -420,7 +411,8 @@ defmodule HTTPoisonTest do
                HTTPoison.get("https://expired.badssl.com/", [], ssl: [{:versions, [:"tlsv1.2"]}])
 
       # Insecure disable verification
-      assert {:ok, _} = HTTPoison.get("https://expired.badssl.com", [], hackney: [:insecure])
+      assert {:ok, _} =
+               HTTPoison.get("https://expired.badssl.com", [], ssl: [verify: :verify_none])
 
       # Can be disabled via verify_fun
       assert {:ok, _} =